moved Error (field) component in shared components
- moved Error.tsx from Pricing
- change paths to Error.tsx in publish
